+++ android/src/oldarch/EspIdfProvisioningSpec.kt
@@ -5,9 +5,9 @@ import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.bridge.ReactContextBaseJavaModule
 
 abstract class EspIdfProvisioningSpec(context: ReactApplicationContext?) : ReactContextBaseJavaModule(context) {
-    abstract fun searchESPDevices(devicePrefix: String, transport: String, security: Int, promise: Promise?)
+    abstract fun searchESPDevices(devicePrefix: String, transport: String, security: Double, promise: Promise?)
     abstract fun stopESPDevicesSearch()
-    abstract fun createESPDevice(deviceName: String, transport: String, security: Int, proofOfPossession: String?, softAPPassword: String?, username: String?, promise: Promise?)
+    abstract fun createESPDevice(deviceName: String, transport: String, security: Double, proofOfPossession: String?, softAPPassword: String?, username: String?, promise: Promise?)
     abstract fun connect(deviceName: String, promise: Promise?)
     abstract fun sendData(deviceName: String, path: String, data: String, promise: Promise?)
     abstract fun scanWifiList(deviceName: String, promise: Promise?)
@@ -22,7 +22,7 @@ abstract class EspIdfProvisioningSpec(context: ReactApplicationContext?) : React
     abstract fun getPrimaryServiceUuid(deviceName: String, promise: Promise?)
     abstract fun setPrimaryServiceUuid(deviceName: String, primaryServiceUuid: String, promise: Promise?)
     abstract fun getSecurityType(deviceName: String, promise: Promise?)
-    abstract fun setSecurityType(deviceName: String, security: Int, promise: Promise?)
+    abstract fun setSecurityType(deviceName: String, security: Double, promise: Promise?)
     abstract fun getTransportType(deviceName: String, promise: Promise?)
     abstract fun getVersionInfo(deviceName: String, promise: Promise?)
     abstract fun getDeviceCapabilities(deviceName: String, promise: Promise?)
